+++ oneflow/user/ops/elu_op.cpp
@@ -21,7 +21,7 @@ namespace {
 
 REGISTER_USER_OP("elu")
     .Input("in")
-    .Attr<double>("alpha")
+    .Attr<float>("alpha")
     .Output("out")
     .SetTensorDescInferFn([](user_op::InferContext* ctx) -> Maybe<void> {
       *ctx->Shape4ArgNameAndIndex("out", 0) = *ctx->Shape4ArgNameAndIndex("in", 0);
@@ -46,7 +46,7 @@ REGISTER_USER_OP("elu")
 REGISTER_USER_OP("elu_grad")
     .Input("x")
     .Input("dy")
-    .Attr<double>("alpha")
+    .Attr<float>("alpha")
     .Output("dx")
     .SetTensorDescInferFn([](user_op::InferContext* ctx) -> Maybe<void> {
       const Shape* x_shape = ctx->Shape4ArgNameAndIndex("x", 0);
@@ -78,7 +78,7 @@ REGISTER_USER_OP_GRAD("elu").SetBackwardOpConfGenFn([](user_op::BackwardOpConfCo
     return builder.OpTypeName("elu_grad")
         .InputBind("x", ctx->FwOp().input("in", 0))
         .InputBind("dy", ctx->FwOp().output_grad("out", 0))
-        .Attr("alpha", ctx->FwOp().attr<double>("alpha"))
+        .Attr("alpha", ctx->FwOp().attr<float>("alpha"))
         .Output("dx")
         .Build();
   });
